Individual News Articles Style
Hopefully this will work in Wordpress fine, especially for content
editing for Annette. Check Foundation’s Document on “Reveal Modal” to
find the specific details on the code. It’s super easy to use. The way
I put the code should be working. The <h2> and <p> content just have to
be edited by Annette.